Разделил утилиты на отдельные файлы
