+++ src/main.js
@@ -3,7 +3,7 @@ import { generatePoint } from './mock/point.js';
 import { offers, destinations } from './mock/const.js';
 
 import { POINT_COUNTER, position, KeyType } from './const.js';
-import { render } from './utils.js';
+import { render } from './utils/render.js';
 
 import SiteHeaderInfoView from './view/info.js';
 import SiteHeaderNavigationView from './view/navigation.js';
