+++ src/mock/point.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 
-import { getRandomInteger } from '../utils.js';
+import { getRandomInteger } from '../utils/common.js';
 import { destinations, offers } from './const.js';
 import { types } from '../const.js';
