+++ src/utils/date.js
@@ -1,5 +1,4 @@
 import dayjs from 'dayjs';
-import { position } from './const.js';
 
 export const dateDiff = (dateFrom, dateTo) => {
   const MINUTE_IN_DAY = 1440;
@@ -37,32 +36,4 @@ export const dateFormat = (date, format = 'YYYY-MM-DD') => {
 
 export const dateToISO = (date) => {
   return dayjs(date).toISOString();
-};
-
-// Получение случайного целого числа
-export const getRandomInteger = (min = 0, max = 1, coefficient = 1) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-
-  return (Math.floor(lower + Math.random() * (upper - lower + 1)) * coefficient);
-};
-
-// Создание элемента
-export const createElement = (template) => {
-  const newElement = document.createElement('div');
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-// Рендер элемента
-export const render = (container, element, place = position.BEFORE_END) => {
-  switch (place) {
-    case position.AFTER_BEGIN:
-      container.prepend(element);
-      break;
-    case position.BEFORE_END:
-      container.append(element);
-      break;
-  }
 };
